roomTitle shown at invites modal

+++ src/pages/OnBoard/OnBoard.tsx
@@ -7,20 +7,31 @@ import {
   ScrollView,
   Modal,
 } from 'react-native';
-import {Button, TextInput, IconButton} from 'react-native-paper';
+import {Button, TextInput, IconButton, DataTable} from 'react-native-paper';
 import useStyles from './OnBoardStyle';
 import {RoomService} from '../../Services/RoomService';
 import {Room} from '../../Models/Room';
 import {ActiveUser, AuthService} from '../../Services/AuthService';
 import {RealTimeService} from '../../Services/RealTimeService';
 function OnBoard({navigation}): JSX.Element {
+  //#region  States
   const [createRoomModalVisible, setCreateRoomModalVisible] = useState(false);
   const [roomNameInput, setroomNameInput] = useState('');
   const [rooms, setRooms] = useState<Room[]>([]);
   const [invites, setInvites] = useState<any[]>([]);
+  const [invitesModalVisible, setInvitesModalVisible] = useState(false);
+  const [invitesPageNumber, setInvitesPageNumber] = React.useState<number>(0);
+  //#endregion
+  //#region ConstVariables
   const styles = useStyles();
   const User = ActiveUser.GetActiveUser();
-
+  const itemPerPage = 5;
+  const fromInvites = invitesPageNumber * itemPerPage;
+  const toInvites = Math.min(
+    (invitesPageNumber + 1) * itemPerPage,
+    invites.length,
+  );
+  //#endregion
   //#region  Service
   const roomService = new RoomService();
   const authService = new AuthService();
@@ -29,12 +40,33 @@ function OnBoard({navigation}): JSX.Element {
 
   useEffect(() => {
     if (User != undefined) {
+      console.log('on1');
       realTimeService.OnInvite(User.user.email).on('child_added', newVal => {
+        console.log('on');
         const inviteId = newVal.val().InviteId;
-        realTimeService.RemoveReadedInvite(inviteId, User.user.email);
-        invites.push(newVal.val());
-        const newInvites = invites.slice();
-        setInvites(newInvites);
+        const roomId = newVal.val().RoomId;
+        console.log(roomId);
+        roomService
+          .GetRoomById(roomId)
+          .then(room => {
+            console.log('under rroom');
+            if (room != undefined) {
+              console.log('defined');
+              console.log(room);
+              realTimeService.RemoveReadedInvite(inviteId, User.user.email);
+              invites.push({
+                InviteId: newVal.val().InviteId,
+                InvitedBy: newVal.val().InvitedBy,
+                Title: room.Title,
+              });
+              const newInvites = invites.slice();
+              setInvites(newInvites);
+            }
+          })
+          .catch(() => {
+            console.log('catch');
+            return;
+          });
       });
     }
 
@@ -89,12 +121,52 @@ function OnBoard({navigation}): JSX.Element {
           </View>
         </View>
       </Modal>
+      <Modal
+        animationType="slide"
+        transparent={true}
+        visible={invitesModalVisible}
+        onRequestClose={() => {
+          setInvitesModalVisible(false);
+        }}>
+        <View style={styles.centeredView}>
+          <View style={styles.modalView}>
+            <IconButton
+              icon="close"
+              size={20}
+              onPress={() => setInvitesModalVisible(false)}
+              style={styles.modalCloseBtn}
+            />
+            <DataTable>
+              <DataTable.Header>
+                <DataTable.Title>Invited By</DataTable.Title>
+                <DataTable.Title>Room Title</DataTable.Title>
+              </DataTable.Header>
+
+              {invites.slice(fromInvites, toInvites).map(item => (
+                <DataTable.Row key={item.InviteId}>
+                  <DataTable.Cell>{item.InvitedBy}</DataTable.Cell>
+                  <DataTable.Cell>{item.Title}</DataTable.Cell>
+                </DataTable.Row>
+              ))}
+
+              <DataTable.Pagination
+                page={invitesPageNumber}
+                numberOfPages={Math.ceil(invites.length / itemPerPage)}
+                onPageChange={page => setInvitesPageNumber(page)}
+                label={`${fromInvites + 1}-${toInvites} of ${invites.length}`}
+                numberOfItemsPerPage={itemPerPage}
+                showFastPaginationControls
+              />
+            </DataTable>
+          </View>
+        </View>
+      </Modal>
       <View style={styles.navbar}>
         <Button
           icon="account-multiple-plus"
           mode="elevated"
           onPress={() => {
-            console.log('join');
+            setInvitesModalVisible(true);
           }}>
           {invites.length > 0 ? `Join (${invites.length})` : 'Join (0)'}
         </Button>
